bot/requirements.txt
websockets>=15.0.1
aiohttp>=3.9.0

# RL and ML libraries
# Note: For GPU acceleration, install with:
# uv pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
stable-baselines3>=2.0.0
gymnasium>=0.29.0
numpy>=1.24.0
matplotlib>=3.7.0
tensorboard>=2.13.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0

# Data handling
pandas>=2.0.0
h5py>=3.9.0

# Logging and monitoring
wandb>=0.15.0
